Home

DynamicEnergyBudgets

DynamicEnergyBudgets — Module.

This is a generalised DEB model. It was developed for plant modelling, but can potentially model any organisms and symbioses.

This model can also be run in microclimates provided by the NicheMapr R package, and can use wide a range of photosynthesis and stomatal conductance formulations from Photosynthesis.jl.

It is also an in-progress attempt at using Julia's multiple-dispatch methods to abstract and generalise DEB theory and maintain a short, maintainable codebase for multiple models - potentially any organism.

Code is adapted from the original DEBtool plant model by Bas Kooijman.

source

Index

Types

DynamicEnergyBudgets.AbstractAllometry

DynamicEnergyBudgets.AbstractAssimilation

DynamicEnergyBudgets.AbstractCarbonAssimilation

DynamicEnergyBudgets.AbstractNH4_NO3Assimilation

DynamicEnergyBudgets.AbstractNitrogenAssimilation

DynamicEnergyBudgets.AbstractScaling

DynamicEnergyBudgets.AbstractStateFeedback

DynamicEnergyBudgets.AbstractTempCorr

DynamicEnergyBudgets.Autophagy

DynamicEnergyBudgets.C3Photosynthesis

DynamicEnergyBudgets.CarbonVars

DynamicEnergyBudgets.KooijmanArea

DynamicEnergyBudgets.KooijmanSLAPhotosynthesis

DynamicEnergyBudgets.Kooijman_NH4_NO3Assimilation

DynamicEnergyBudgets.Maturity

DynamicEnergyBudgets.N_Assimilation

DynamicEnergyBudgets.NitrogenVars

DynamicEnergyBudgets.Organ

DynamicEnergyBudgets.Organism

DynamicEnergyBudgets.Params

DynamicEnergyBudgets.Records

DynamicEnergyBudgets.Scenario

DynamicEnergyBudgets.SharedParams

DynamicEnergyBudgets.TempCorr

DynamicEnergyBudgets.TempCorrLower

DynamicEnergyBudgets.TempCorrLowerUpper

DynamicEnergyBudgets.Vars

Functions

DynamicEnergyBudgets.allometric_height

DynamicEnergyBudgets.assimilation!

DynamicEnergyBudgets.assimilation!

DynamicEnergyBudgets.assimilation!

DynamicEnergyBudgets.assimilation!

DynamicEnergyBudgets.catabolic_fluxes

DynamicEnergyBudgets.catabolism!

DynamicEnergyBudgets.correct_temps!

DynamicEnergyBudgets.debmodel!

DynamicEnergyBudgets.dissipation!

DynamicEnergyBudgets.feedback!

DynamicEnergyBudgets.find_rate

DynamicEnergyBudgets.germinated

DynamicEnergyBudgets.growth!

DynamicEnergyBudgets.half_saturation

DynamicEnergyBudgets.maintenence!

DynamicEnergyBudgets.maturity!

DynamicEnergyBudgets.metabolism!

DynamicEnergyBudgets.photosynthesis

DynamicEnergyBudgets.photosynthesis

DynamicEnergyBudgets.product!

DynamicEnergyBudgets.rate_formula

DynamicEnergyBudgets.reserve_drain!

DynamicEnergyBudgets.reserve_loss!

DynamicEnergyBudgets.reuse_rejected!

DynamicEnergyBudgets.runmodel!

DynamicEnergyBudgets.runmodel!

DynamicEnergyBudgets.scaling

DynamicEnergyBudgets.set_cur_records!

DynamicEnergyBudgets.set_cur_records!

DynamicEnergyBudgets.setstate!

DynamicEnergyBudgets.stoich_merge

DynamicEnergyBudgets.sumflux!

DynamicEnergyBudgets.synthesizing_unit

DynamicEnergyBudgets.tempcorr

DynamicEnergyBudgets.translocate!

DynamicEnergyBudgets.translocation!

DynamicEnergyBudgets.uptake_nitrogen

DynamicEnergyBudgets.uptake_nitrogen

DynamicEnergyBudgets.κtra

Parameters and Types

DynamicEnergyBudgets uses the julia type system to organise both the parameters and the logic of the model. This means the model is completely modular and customisable.

DynamicEnergyBudgets.AbstractAssimilation — Type.

abstract AbstractAssimilation

Assimilation

source

DynamicEnergyBudgets.AbstractCarbonAssimilation — Type.

abstract AbstractCarbonAssimilation <: DynamicEnergyBudgets.AbstractAssimilation

Paraent of all Carbon assimilation types

source

DynamicEnergyBudgets.AbstractNitrogenAssimilation — Type.

abstract AbstractNitrogenAssimilation <: DynamicEnergyBudgets.AbstractAssimilation

Paraent of all Nitrogen assimilation types

source

DynamicEnergyBudgets.AbstractScaling — Type.

abstract AbstractScaling

Surface area scaling rules

source

DynamicEnergyBudgets.AbstractStateFeedback — Type.

abstract AbstractStateFeedback

State feedback parameters. These modfy state based on state.

source

DynamicEnergyBudgets.AbstractTempCorr — Type.

abstract AbstractTempCorr

Temperature correction parameters

source

DynamicEnergyBudgets.Autophagy — Type.

Autophagy. Parameters for self reabsorbtion when metabolic rates fall

source

DynamicEnergyBudgets.C3Photosynthesis — Type.

Uses FvCB photosynthesis model from Photosynthesis.jl

source

DynamicEnergyBudgets.CarbonVars — Type.

Variables for carbon assimilation

source

DynamicEnergyBudgets.KooijmanArea — Type.

Surface areai scaling curve. Simulates growth and shade crowding later in life.

source

DynamicEnergyBudgets.KooijmanSLAPhotosynthesis — Type.

Parameters for simple photosynthesis module. With specific leaf area to convert area to mass

source

DynamicEnergyBudgets.Maturity — Type.

Maturity parameters. Seperated to make maturity modeling optional, reducing complexity

source

DynamicEnergyBudgets.N_Assimilation — Type.

Parameters for lumped Nitrogen assimilation

source

DynamicEnergyBudgets.NitrogenVars — Type.

Variables for nitgroen assimilation

source

DynamicEnergyBudgets.Organ — Type.

type Organ{S, P, SH, V, F, F1}

Basic model components. For a plants, organs might be roots, shoots and leaves

source

DynamicEnergyBudgets.Organism — Type.

struct Organism{R, N, P}

A single organism. Models can be run directly for an organism without an environment.

source

DynamicEnergyBudgets.Params — Type.

Model parameters that vary between organs

source

DynamicEnergyBudgets.Scenario — Type.

A scenario is an organism (or potentially organisms) growing in response to real or simulated environmental data.

source

DynamicEnergyBudgets.SharedParams — Type.

Model parameters shared between organs

source

DynamicEnergyBudgets.TempCorr — Type.

Simple temperature correction parameters

source

DynamicEnergyBudgets.TempCorrLower — Type.

Temperature correction with lower boudn parameters

source

DynamicEnergyBudgets.TempCorrLowerUpper — Type.

Temperature correction with lower and upper bound parameters

source

DynamicEnergyBudgets.Vars — Type.

Model variables

source

DynamicEnergyBudgets.AbstractAllometry — Type.

abstract AbstractAllometry

Allometry. Scaling rules to relate size to mass.

source

DynamicEnergyBudgets.AbstractNH4_NO3Assimilation — Type.

abstract AbstractNH4_NO3Assimilation <: DynamicEnergyBudgets.AbstractNitrogenAssimilation

Parent of Ammonia/Nitrate assimilation types

source

DynamicEnergyBudgets.Kooijman_NH4_NO3Assimilation — Type.

Parameters for Ammonia/Nitrate assimilation

source

DynamicEnergyBudgets.Records — Type.

struct Records{V, F, F1}

Records of variables and flux for ploting and analysis

source

Model

Core functions of the model.

DynamicEnergyBudgets.debmodel! — Method.

A generalised multi-reserve, multi-organ Dynamic Energy Budget model.

Applies metabolism, translocation and assimilation mehtods to N organs.

settings is a struct with required model data, DEBSettings or similar. t is the timestep

source

DynamicEnergyBudgets.runmodel! — Method.

Method to run a DEB organism model, with a static environment.

source

DynamicEnergyBudgets.runmodel! — Method.

Method to run a DEB scenario, with environment applied to organism(s).

source

DynamicEnergyBudgets.allometric_height — Method.

source

DynamicEnergyBudgets.catabolism! — Method.

catabolism!(o, u::AbstractStateCNE, t::Number)

Catabolism for E, C and N, or C, N and E reserves. Does not finalise flux in J - operates only on J1 (intermediate storage)

source

DynamicEnergyBudgets.dissipation! — Method.

Dissipation for any reserve. Growth, maturity and maintenence are grouped as dissipative processes.

source

DynamicEnergyBudgets.feedback! — Method.

Function to apply feedback on growth the process, such as autopagy in resource shortage.

Without a function like this you will likely be occasionally breaking the laws of thermodynamics by introducing negative rates.

source

DynamicEnergyBudgets.find_rate — Method.

Calculate rate formula. TODO: use Roots.jl for this

source

DynamicEnergyBudgets.germinated — Method.

Check if germination has happened. Independent for each organ, although this may not make sense.

source

DynamicEnergyBudgets.growth! — Method.

Allocates reserves to growth.

source

DynamicEnergyBudgets.maintenence! — Method.

Allocates reserve drain due to maintenance.

source

DynamicEnergyBudgets.maturity! — Method.

maturity!(f, o, u)

Allocates reserve drain due to maturity maintenance. Stores in M state variable if it exists.

source

DynamicEnergyBudgets.metabolism! — Method.

Metabolism is an identical process for all organs, with potentially different parameters or area and rate functions.

source

DynamicEnergyBudgets.product! — Method.

Allocates waste products from growth and maintenance.

source

DynamicEnergyBudgets.reserve_drain! — Method.

Generalised reserve drain for any flux column col (ie :gro) and any combination of reserves.

source

DynamicEnergyBudgets.reserve_loss! — Method.

Generalised reserve loss to track carbon.

source

DynamicEnergyBudgets.reuse_rejected! — Method.

Reallocate state rejected from synthesizing units.

TODO add a 1-organs method Also how does this interact with assimilation?

source

DynamicEnergyBudgets.scaling — Method.

source

DynamicEnergyBudgets.translocate! — Method.

Versions for E, CN and CNE reserves.

Translocation is occurs between adjacent organs. This function is identical both directiono, so on represents whichever is not the current organs. Will not run with less than 2 organs.

FIXME this will be broken for organs > 2

source

DynamicEnergyBudgets.translocation! — Method.

Versions for E, CN and CNE reserves.

Translocation is occurs between adjacent organs. This function is identical both directiono, so on represents whichever is not the current organs. Will not run with less than 2 organs.

FIXME this will be broken for organs > 2

source

DynamicEnergyBudgets.κtra — Method.

κtra is the difference paramsbetween κsoma and κrep

source

Environment

DynamicEnergyBudgets.correct_temps! — Method.

Scale variables by temperature

source

Assimilation

Carbon and nitrogen assimilation functions. These are the main points of difference between organs.

DynamicEnergyBudgets.assimilation! — Method.

assimilation!(o1, o2)

Runs assimilation methods, depending on formulation and state.

source

DynamicEnergyBudgets.assimilation! — Method.

assimilation!(f::AbstractCarbonAssimilation, o1, o2, u::AbstractStateCNE)

Runs nitrogen uptake, and combines N with translocated C.

source

DynamicEnergyBudgets.assimilation! — Method.

assimilation!(f::AbstractNH4_NO3Assimilation, o1, o2, u::AbstractStateCNE)

Runs nitrogen uptake for nitrate and ammonia, and combines N with translocated C. Unused ammonia is discarded.

source

DynamicEnergyBudgets.assimilation! — Method.

assimilation!(f::N_Assimilation, o1, o2, u::AbstractStateCNE)

Runs nitrogen uptake, and combines N with translocated C.

source

DynamicEnergyBudgets.photosynthesis — Method.

photosynthesis(f::C3Photosynthesis, o1, o2)

Returns carbon assimilated in mols per time.

source

DynamicEnergyBudgets.photosynthesis — Method.

photosynthesis(f::KooijmanSLAPhotosynthesis, o1, o2)

Returns carbon assimilated in mols per time.

source

DynamicEnergyBudgets.uptake_nitrogen — Method.

uptake_nitrogen(f::Kooijman_NH4_NO3Assimilation, o1, o2)

Returns total nitrogen, nitrate and ammonia assimilated in mols per time.

source

DynamicEnergyBudgets.uptake_nitrogen — Method.

uptake_nitrogen(f::N_Assimilation, o1, o2)

Returns nitrogen assimilated in mols per time.

source

Functions

Low-level DEB theory functions

DynamicEnergyBudgets.catabolic_fluxes — Method.

catabolic_fluxes(ureserve, A_turnover, r)

Returns the current catabolic flux at rate r, or the flux as a proportion of u[V], depending on ureserve values.

source

DynamicEnergyBudgets.half_saturation — Method.

half_saturation(max, half, x)

Half satration curve.

source

DynamicEnergyBudgets.rate_formula — Method.

rate_formula(r, ureserve::NTuple, A_turnover::NTuple, j_E_mai, y_V_E, κsoma)

Rate formulas for E, CN or CNE reserves

source

DynamicEnergyBudgets.stoich_merge — Method.

stoich_merge(a, b)

Merge two inputs stoichiometrically. The minimum value is limiting, and stochasticity of pairing is simulated so that for any a, b stoich_merge(a, b) < a, to the limit b → ∞ where stoich_merge(a, b) = a

source

DynamicEnergyBudgets.synthesizing_unit — Method.

synthesizing_unit(Ja, Jb, ya, yb)

Merge fluxes stoichiometrically into general reserve Eab based on yeild fractions ya and yb. The remainder is returned as unmixed reserves Ea and Eb.

source

DynamicEnergyBudgets.tempcorr — Method.

tempcorr(T, T1, A, [L, AL,] [H, AH])

DEB tempcorr function. Uses lower and uppper bounds if they are supplied. Temperatures all in Kelvins.

source

Setup and utilities

DynamicEnergyBudgets.set_cur_records! — Method.

update vars and flux to record for time t

source

DynamicEnergyBudgets.set_cur_records! — Method.

Void records so we just stay on the same record

source

DynamicEnergyBudgets.setstate! — Method.

copy the diffeq state to organs

source

DynamicEnergyBudgets.sumflux! — Method.

sum flux matrix

source