DynamicEnergyBudgets

DynamicEnergyBudgets.DynamicEnergyBudgets — Module

This package is a generalised DEB model. It was developed for plant modelling, but can potentially be used to model any organisms and symbioses.

This model can also be run in microclimates provided by the NicheMapr R package, and can use wide a range of photosynthesis and stomatal conductance formulations from Photosynthesis.jl.

It is also an in-progress attempt at using Julia's multiple-dispatch methods to abstract and generalise DEB theory and maintain a short, maintainable codebase for multiplt models - potentially any organism. Code is adapted from the original DEBtool plant model by Bas Kooijman.

Model

Core functions of the model.

DynamicEnergyBudgets.AbstractOrgan — Type

abstract type AbstractOrgan

Abstract supertype for organs. Inherit from it if you need to difine behaviour diferent to that or Organ.

DynamicEnergyBudgets.AbstractParams — Type

abstract type AbstractParams

Astract supertype for organs parameters. Extend to change the components that are specific to each organ.

DynamicEnergyBudgets.Organ — Type

struct Organ{P, S, V, F} <: AbstractOrgan{P,S}

Organ(params, shared, vars, J)

Basic model components. For a plants, organs might be roots, stem and leaves

Field	Description	Default	Bounds

params		nothing	(0.0, 1.0)

shared		nothing	(0.0, 1.0)

vars		nothing	(0.0, 1.0)

J		nothing	(0.0, 1.0)

DynamicEnergyBudgets.Organ — Method

Organ(params, shared, records)

Construct an organ from parameters, shared parameters and views into records arrays for vaiable and flux matrices.

DynamicEnergyBudgets.Plant — Method

Plant(states=(:V, :C, :N),
      transformations=(:asi, :gro, :mai, :rej, :res),
      catstates=(:CN, :C, :N, :E),
      cattransformations=(:ctb,),
      params=(ShootParamsCN(), RootParamsCN()),
      vars=(Vars(), Vars()),
      shared=SharedParams(),
      records=nothing,
      environment=nothing,
      time=0.0hr:1.0hr:8760.0hr,
      environment_start=Ref(1.0hr),
      dead=Ref(false))

Outer construtor for defaults

DynamicEnergyBudgets.Plant — Method

(organism::Plant)(du, u, p, t::Number)

The Plant struct can be used as a function to run the DEB model in a solver, such as (i.e. from DiffEq).

du is the change in state vector written to by the model, u is the vector of current state, p is a tuple or vector of parameter values, and t is time.

If p is not nothing the model will be rebuilt with the new parameters. Note that this works with any model combination automatically as long as the number of parameters matches the number of parameters in the model.

DynamicEnergyBudgets.debmodel! — Method

debmodel!(organs::Tuple, u::Tuple{AbstractArray}, env)

A generalised multi-reserve, multi-organ Dynamic Energy Budget model.

Applies metabolism, translocation and assimilation methods to 2 organs.

settings is a struct with required model data, DEBSettings or similar. t is the timestep

DynamicEnergyBudgets.AbstractSharedParams — Type

abstract type AbstractSharedParams

Astract supertype for shared parameters. Extend to change the components that are shared.

DynamicEnergyBudgets.AbstractVars — Type

abstract type AbstractVars

Model variables. Allow storing and accessing variables for use by multiple components.

DynamicEnergyBudgets.define_organs — Method

define_organs(o::AbstractOrganism, t)

Organs are constructed with views of Records and J Arrays at time t

DynamicEnergyBudgets.metabolism! — Method

metabolism!(organs::Tuple, u)

Metabolism is an identical process for all organs, with potentially different parameters or area and rate functions.

Core Parameters

DynamicEnergyBudgets.growth! — Method

growth!(o, u)

Allocates reserves to growth flux, generalised for any number of reserves.

DynamicEnergyBudgets.maintenence! — Method

Allocates reserve drain due to maintenance, generalised for any number of reserves.

Maintenance is temperature dependent.

Components

Allometry

DynamicEnergyBudgets.AbstractAllometry — Type

abstract type AbstractAllometry

Allometry. Scaling rules to relate size to mass.

Assimilation

DynamicEnergyBudgets.AbstractCAssim — Type

abstract type AbstractCAssim <: DynamicEnergyBudgets.AbstractAssim

Parent of all Carbon assimilation types

DynamicEnergyBudgets.AbstractNAssim — Type

abstract type AbstractNAssim <: DynamicEnergyBudgets.AbstractAssim

Parent of all Nitrogen assimilation types

DynamicEnergyBudgets.AbstractAssim — Type

abstract type AbstractAssim

Assimilation

DynamicEnergyBudgets.AbstractNH4_NO3Assim — Type

abstract type AbstractNH4_NO3Assim <: AbstractNAssim

Parent of Ammonia/Nitrate assimilation types

Catabolism

Environment

DynamicEnergyBudgets.ManualTemperature — Type

struct ManualTemperature{T}

ManualTemperature(airtemperature, soiltemperature)

Environment for simple manual temperature control.

Field	Description	Default	Bounds

airtemperature		nothing	(0.0, 1.0)

soiltemperature		nothing	(0.0, 1.0)

Maintenance

Maturity

Rate

Resorption

DynamicEnergyBudgets.AbstractResorption — Type

abstract type AbstractResorption

Resorption. Parameters for reabsorbtion of nutrients from structures when metabolic rates fall.

Shape

Synthesizing Units

DynamicEnergyBudgets.MinimumRuleSU — Type

struct MinimumRuleSU <: AbstractSynthesizingUnit

MinimumRuleSU(k)

0-parameter synthesizing unit where law of the minimum controls the production of one compound form two other compounds.

DynamicEnergyBudgets.ParallelComplementarySU — Type

struct ParallelComplementarySU <: AbstractSynthesizingUnit

ParallelComplementarySU(k)

0-parameter synthesizing unita.

Temperature Correction

DynamicEnergyBudgets.AbstractTemperatureCorrection — Type

abstract type AbstractTemperatureCorrection

Temperature correction parameters

Translocation

DynamicEnergyBudgets.LosslessPassiveTranslocation — Type

struct LosslessPassiveTranslocation <: PassiveTranslocation

LosslessPassiveTranslocation)

Parameterless rejection where substrate rejected from synthesizing units during catabolism is returned to reserve without loss.

Other functions

Low-level DEB theory functions

DynamicEnergyBudgets.half_saturation — Method

half_saturation(max, half, x)

Half satration curve.

DynamicEnergyBudgets.reserve_drain! — Method

reserve_drain!(o::AbstractOrgan, column, drain)

Generalised reserve drain for any flux column (ie :gro), specified by its Symbol or Int index.

DynamicEnergyBudgets.sum_flux! — Method

sum_flux!(du, organs::Tuple)

Sum flux matrix and write to du.

DynamicEnergyBudgets.check_params — Method

check_params(o)

Check DEB parameters to avoid breaking mass balance. Can be pased an organ or tuple of organs.

DynamicEnergyBudgets.split_state — Method

split_state(o::Tuple, u::AbstractArray)

Split state vector into multiple views, one for each organ.

We do this recursively for type stability.